Add support for eth_call() in Seth RPC (STL-1568)
The model here is straightforward:
 - Add a read-only message call
 - Fix the machinery for passing block_id to Burrow
 - Fix the machinery for TAG|QUANTITY elements
 - Send the transaction the normal way then poll it out

Includes a handful of drive-by fixes to minor errors and build problems
hand-ported from upstream.  These should be benign (and were required for the
copmonent to build against changing dependencies) but I make this note to avoid
surprising reviewers, especially as it's all in one big commit.

Signed-off-by: JAG-UK <jon.geater@gmail.com>